event controller to database done

+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/controllers/OrgController.java
@@ -3,18 +3,23 @@ package hu.project.groupproject.resourceserver.controllers;
 import org.springframework.web.bind.annotation.RestController;
 
 import hu.project.groupproject.resourceserver.dtos.ImageUploadDetailsDto;
+import hu.project.groupproject.resourceserver.dtos.En.EventDtoPublic;
 import hu.project.groupproject.resourceserver.dtos.En.orgs.OrgDtoCreate;
 import hu.project.groupproject.resourceserver.dtos.En.orgs.OrgDtoPublic;
+import hu.project.groupproject.resourceserver.dtos.En.posts.out.PostDtoPublicExtended;
 import hu.project.groupproject.resourceserver.entities.softdeletable.MyUser;
+import hu.project.groupproject.resourceserver.services.EventService;
 import hu.project.groupproject.resourceserver.services.OrgService;
 
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 
 import javax.management.InvalidAttributeValueException;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -34,29 +39,35 @@ public class OrgController {
     protected final Log logger = LogFactory.getLog(getClass());
 
     OrgService orgService;
+    EventService eventService;
 
-    public OrgController(OrgService orgService){
+    public OrgController(OrgService orgService, EventService eventService){
         this.orgService=orgService;
+        this.eventService=eventService;
     }
     
     @PostMapping("/addAdmin")
     @PreAuthorize("hasAnyRole('ADMIN','ORG_ADMIN')")
-    public void addUserToOrg(@RequestBody Map<String, String> body){
+    public void addUserToOrg(@RequestBody Map<String, String> body, Authentication authentication){
+        MyUser user = (MyUser)authentication.getPrincipal();
         String adminId = body.get("adminId");
-        String userId = body.get("userId");
-        String orgId = body.get("orgId");
-        
-        orgService.addUser(orgId,adminId,userId);
+        if(user.getId()==adminId){
+            String userId = body.get("userId");
+            String orgId = body.get("orgId");
+            orgService.addUser(orgId,adminId,userId);
+        }
     }
     //should be more sophisticated so not any admin can remove any and all other admins...
     @PostMapping("/removeAdmin")
-    @PreAuthorize("hasRole('ADMIN')")
-    public void removeUserFromOrg(@RequestBody Map<String, String> body){
-        String adminId =body.get("adminId");
-        String userId =body.get("userId");
-        String orgId =body.get("orgId");
-        
-        orgService.removeUser(orgId,adminId,userId);
+    @PreAuthorize("hasRole('ORG_ADMIN')")
+    public void removeUserFromOrg(@RequestBody Map<String, String> body, Authentication authentication){
+        MyUser user = (MyUser)authentication.getPrincipal();
+        String adminId = body.get("adminId");
+        if(user.getId()==adminId){
+            String userId = body.get("userId");
+            String orgId = body.get("orgId");
+            orgService.removeUser(orgId,adminId,userId);
+        }
     }
     
     @PostMapping("/new")
@@ -67,8 +78,12 @@ public class OrgController {
     
     @PutMapping("/{id}")
     @PreAuthorize("hasRole('ORG_ADMIN')")
-    public ImageUploadDetailsDto updateOrg(@PathVariable String orgId, @RequestBody OrgDtoCreate org) throws InvalidAttributeValueException{
-        return orgService.saveOrg(orgId,org);
+    public ImageUploadDetailsDto updateOrg(@PathVariable String orgId, @RequestBody OrgDtoCreate org, Authentication authentication) throws InvalidAttributeValueException{
+        MyUser user = (MyUser)authentication.getPrincipal();
+        if (org.adminId()==user.getId()) {
+            return orgService.saveOrg(orgId,org);
+        }
+        throw new AccessDeniedException("You don't have the right to change this organisation");
     }
     
     @GetMapping("/{id}")
@@ -82,6 +97,11 @@ public class OrgController {
         MyUser user = (MyUser)auth.getPrincipal();
         logger.debug(user.getId());
         orgService.deleteOrg(user.getId(),orgId);
+    }
+
+    @GetMapping("/{orgId}/events")
+    public Set<EventDtoPublic> getPostsForUser(@PathVariable String orgId) {
+        return eventService.getEventsForOrg(orgId);
     }
     
 }
